cinema-app/cinema/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
djangorestframework==3.14.0
exceptiongroup==1.0.1
Faker==15.2.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tomli==2.0.1
pymongo==3.12.1
djongo==1.3.6
django_environ==0.9.0
